project-cluster/requirements.txt
# Core Dependencies
pandas>=1.3.0
numpy>=1.21.0
PyYAML>=5.4.0

# Machine Learning and Data Analysis
scikit-learn>=0.24.0
scipy>=1.7.0

# Visualization
matplotlib>=3.4.0

# Optional: For improved performance
# numexpr>=2.7.0
# bottleneck>=1.3.0